@@ -22,28 +22,29 @@
     }
   },
   "dependencies": {
-    "@nuxtjs/pwa": "^3.3.5",
-    "core-js": "^3.9.1",
-    "date-fns": "^2.20.0",
-    "nuxt3": "latest"
+    "@kevinmarrec/nuxt-pwa": "^0.4.2",
+    "core-js": "^3.25.3",
+    "date-fns": "^2.29.3",
+    "nuxt": "^3.0.0-rc.11"
   },
   "devDependencies": {
-    "@vue/test-utils": "^1.1.3",
-    "eslint": "^7.22.0",
-    "eslint-config-prettier": "^8.1.0",
-    "eslint-plugin-nuxt": "^2.0.0",
-    "eslint-plugin-prettier": "^3.3.1",
-    "eslint-plugin-vue": "^7.7.0",
-    "husky": "^4.3.8",
-    "jest": "^26.6.3",
-    "lint-staged": "^10.5.4",
-    "prettier": "^2.2.1",
-    "sass": "^1.50.0",
-    "sass-loader": "^10.1.1",
-    "stylelint": "^13.12.0",
-    "stylelint-config-prettier": "^8.0.2",
-    "stylelint-config-standard": "^21.0.0",
-    "ts-jest": "^26.5.4",
-    "vue-jest": "^3.0.4"
+    "@nuxtjs/eslint-config-typescript": "^11.0.0",
+    "@vue/test-utils": "^2.0.2",
+    "eslint": "^8.24.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-nuxt": "^4.0.0",
+    "eslint-plugin-prettier": "^4.2.1",
+    "eslint-plugin-vue": "^9.5.1",
+    "husky": "^8.0.1",
+    "jest": "^29.0.3",
+    "lint-staged": "^13.0.3",
+    "prettier": "^2.7.1",
+    "sass": "^1.55.0",
+    "sass-loader": "^13.0.2",
+    "stylelint": "^14.12.1",
+    "stylelint-config-prettier": "^9.0.3",
+    "stylelint-config-standard": "^28.0.0",
+    "ts-jest": "^29.0.2",
+    "vue-jest": "^3.0.7"
   }
 }
